PRSD-835: Moves emailModels into viewModels
