+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/emailModels/LandlordRegistrationConfirmationEmail.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.emailModels
+package uk.gov.communities.prsdb.webapp.models.viewModels.emailModels
 
 data class LandlordRegistrationConfirmationEmail(
     val lrn: String,
